reorganize test data files and fixtures
