+++ tests/conftest.py
@@ -1,7 +1,13 @@
+import warnings
+from pathlib import Path
+
+import mudata as md
 import numpy as np
 import pandas as pd
 import pytest
 from anndata import AnnData
+
+import prismo as pr
 
 
 @pytest.fixture(scope="module")
@@ -54,3 +60,20 @@ def random_adata(rng, random_array):
         return adata
 
     return _adata
+
+
+@pytest.fixture(scope="session")
+def cll_data():
+    with warnings.catch_warnings():
+        warnings.simplefilter("ignore", category=FutureWarning)
+        return md.read_h5mu(Path(__file__).parent / "data" / "cll.h5mu")
+
+
+@pytest.fixture(scope="session")
+def cll_model():
+    return pr.PRISMO.load(Path(__file__).parent / "data" / "cll_model.h5")
+
+
+@pytest.fixture(scope="session")
+def mousebrain_model():
+    return pr.PRISMO.load(Path(__file__).parent / "data" / "mousebrain_model.h5")
