+++ tests/data/mousebrain_model.h5
